Creates database first script and changes query id type from long to int. The idea is to the project to be self hosted so there's no need to have such a big index count.

+++ Backoffice.Adapters.QueryRunner/Adapters/QueryRunner.cs
@@ -7,7 +7,7 @@ namespace Backoffice.Adapters.QueryRunner.Adapters
 {
     public class QueryRunner : IQueryRunner
     {
-        public Task<Result> Run(long id)
+        public Task<Result> Run(int id)
         {
             throw new NotImplementedException();
         }
